pyproject.toml
[project]
name = "vigil-contract"
version = "0.1.0"
description = "Hyperliquid Trader Intelligence Engine"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Existing
    "boto3>=1.35.0",
    "ipykernel>=6.0.0",
    "lz4>=4.0.0",
    "msgpack>=1.0.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "tqdm>=4.66.0",

    # Data processing
    "polars>=1.0.0",
    "pyarrow>=14.0.0",

    # Database
    "psycopg[binary]>=3.1.0",

    # Data science
    "scikit-learn>=1.3.0",
    "umap-learn>=0.5.0",

    # Visualization
    "plotly>=5.18.0",
    "seaborn>=0.13.0",
    "matplotlib>=3.8.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
